+++ src/apycfixedarray_wrapper.cc
@@ -238,16 +238,29 @@ void bind_cfixed_array(nb::module_& m)
             )pbdoc"
         )
 
-        .def("to_numpy", &APyCFixedArray::to_numpy, R"pbdoc(
+        .def(
+            "to_numpy",
+            &APyCFixedArray::to_numpy,
+            nb::arg("dtype") = nb::none(),
+            nb::arg("copy") = nb::none(),
+            R"pbdoc(
             Return array as a :class:`numpy.ndarray` of :class:`numpy.complex128`.
 
             The returned array has the same `shape` and values as `self`. This
             method rounds away from infinity on ties.
 
+            Parameters
+            ----------
+            dtype : :std:doc:`numpy:dtype`
+                The desired data type of the output array. This parameter is currently ignored.
+            copy : :class:`bool`
+                Whether to copy the data or not. Must be :code:`True` or :code:`None`.
+
             Returns
             -------
             :class:`numpy.ndarray`
-            )pbdoc")
+            )pbdoc"
+        )
 
         //.def(
         //    "to_bits",
@@ -1222,7 +1235,12 @@ void bind_cfixed_array(nb::module_& m)
         //             );
         //         }
         //     )
-        .def("__array__", &APyCFixedArray::to_numpy)
+        .def(
+            "__array__",
+            &APyCFixedArray::to_numpy,
+            nb::arg("dtype") = nb::none(),
+            nb::arg("copy") = nb::none()
+        )
 
         ;